magit-log-edit-commit: use magit-mode-get-buffer-create
In `magit-log-edit-commit' use `magit-mode-get-buffer-create' instead of
`magit-find-buffer' which has been removed from `magit.el'.  Note that
we don't switch to simply `magit-mode-get-buffer'.  That, as well as the
previously used `magit-find-buffer' assume that the status buffer exits,
which isn't necessarily true, so create it if it doesn't exist.

(The correct thing to do would be not to expect that buffer to exist at
all, but remember that this package is in limbo - I have not interest in
adding any new features and only make adjustments if doing so becomes
necessary because of changes in `magit.el'.)